+++ src/database/queries/changemakerFieldValueBatches/insertOne.sql
@@ -1,9 +1,11 @@
 INSERT INTO changemaker_field_value_batches (
 	source_id,
-	notes
+	notes,
+	created_by
 ) VALUES (
 	:sourceId,
-	:notes
+	:notes,
+	:authContextKeycloakUserId
 )
 RETURNING changemaker_field_value_batch_to_json(changemaker_field_value_batches)
 	AS object;